Allow changing of command workdir
Often when we download a tarball we then need to cd in to run commands. This is both repetetive and hard to do in our current structure; the double ampersand, for example, being a shell construct and not an exec construct, wont work and enforcing commands for everything is just ugly and introduces work.