Ensure tracking operation progress ease
Prior to this change, the output from vind (via vin) was reasonbly confusing; there was no real explanation of what line belonged to what. Come back in the middle of an install and you've no idea what package the output on the screen belongs to. This commit introduces a bright cyan-bold package at the start of every line