Allow some manifests to avoid polluting env
Some manifests, such as m4, fail because setting stuff like CFLAGS
requires autoconf to be run, which doesn't exist. Installing autoconf is
a pain because of some f-ing awful circular dependencies between
autoconf, automake, binutils, and texutils.

So rather than fight that bullshirt, we need to tell m4 to just build as a
completely vanilla `./configure && make && make install` package. The
best way of doing that is to avoid setting *anything* in the environment
at all, and letting things be.

This may become useful elsewhere, too.