Install multiple packages together as a meta package
Because we now support triggers, and because triggers need to run at the absolute end, and because the potential install loop is a pain, we can install build a meta package when many packages are set. So long as that package stays out of the state db, so much the better.

See: https://github.com/vinyl-linux/vin/issues/44